chore: refactor form step components
Refactored the form by creating two components for the form steps. Step one includes options to send the ID of the game and a button to advance. Step two displays the image, name, and price of the game, which the user can modify. Added jump-in and jump-out animations for the form. Created new global interfaces for better code organization.

Author: Wellesley Mussolini