feat: project route sidebar enum
This enum acts as a central hub for all the routes used in the project's sidebar, covering both the ones everyone can see and the private ones. Its main goal is to avoid the hassle of rewriting the same route names across different parts of the project. Instead of scattering route names all over the codebase, this enum puts them all in one place. This way, if you ever need to update a route, you do it once in the enum, and it’s good to go everywhere it’s used. It makes the code cleaner and your life easier, especially when the project grows and you need to add more routes. It’s a practical solution that keeps the route management straightforward.