features the sidebar's new settings  for the users with 'ADMIN' role
The useFilterSidebar function has been updated to streamline the organization and display of sidebar navigation elements across our platform. This hook is designed to manage two key arrays that hold the routes, icons, and labels for the sidebar items. The first array caters to the default user experience, providing navigation options that are visible and accessible to all users. This includes standard routes with their respective icons and labels, ensuring a clear and consistent navigation experience for everyone. The second array is specifically designed for administrators, containing items that appear under the /admin route. This admin-specific navigation setup includes routes and tools necessary for administrative tasks, helping to segregate more sensitive functionalities from the general user interface. By using this structured approach, the useFilterSidebar ensures that the sidebar remains organized and tailored to user roles, facilitating a cleaner codebase and easier updates to the sidebar components as the project's needs evolve.

Author: Wellesley Mussolini