introduces the sidebar's user account menu
The account menu is located at the bottom of the sidebar and is designed to interact seamlessly with the sidebar's expandable feature. This part of the sidebar displays the user's profile icon, name, and email, along with the logout component. It's responsive to the state of the sidebar: when the sidebar is expanded (a boolean state set to true), it shows both the icons and labels; when it's collapsed (state set to false), it hides the labels and only displays the icons. This dynamic adjustment ensures that the user interface remains clean and intuitive, regardless of the sidebar's state, maintaining usability while offering a compact view when needed. This component acts as the functional footer of the sidebar, enhancing the user's navigation experience by adapting to the expandable and retractable nature of the sidebar.

Author: Wellesley Mussolini