refactor: relocate auth options to dedicated (auth) folder and improve session handling
The uthOptions function was moved from the global utils folder to the utils folder inside the uth folder, located in src/app/(auth).
This change simplifies understanding and managing the authentication structure, making debugging and further modifications easier.

In addition to the relocation, updates were made to the uthOptions. The session now uses the strategy: 'jwt', enabling JWT-based sessions. A new jwt method was introduced to set the token’s role based on the user session. The method checks if a user session exists, and if so, assigns the token's role to the user's role before returning the token.

I also commented out the previous cookie logic in the session method and now directly return the session values necessary for the application. The session secret is now provided by the .env file, referencing the OAuth secret for Google authentication. This modification allows the getToken method from NextAuth to retrieve the user session directly in the middleware, removing the need for a complex cookie logic that previously set the user’s role.

These changes not only reduce redundant code but also simplify the process of checking user roles and blocking routes based on authentication status, streamlining the overall logic.