refactor: optimize middleware by removing cookie-based checks for authentication
With the recent changes in the uthOptions, the middleware no longer relies on cookies to check the user's authentication status or role. Now, the getToken function retrieves the session directly, allowing us to access the user's role and authentication status without the need for the previous cookie logic.

This refactor significantly reduces the amount of code in the middleware by removing unnecessary cookie checks, variables, and functions. Instead, checks and conditions are now based on the getToken session. This enables more efficient route handling, such as blocking or allowing access based on whether the user is authenticated and their session role. The overall result is a cleaner and more streamlined middleware.