docs: update notes file with modifications to reduce redundancy
Updated the notes file to document modifications aimed at reducing redundancy in the file handling database requests. Noted that two files are creating conflicts: the API file with HTTP methods and the Prisma method directly to the database. Plan to remove the API file and write the HTTP methods in the BoardGame service that currently is handling the logic directly through Prisma.