refactor code and clean up unnecessary lines
I went through the code and removed a bunch of lines that I thought I might need later but have realized they're not necessary for the current implementation. It was just cluttering things up, so I decided to clean it out. I also wrapped up a few tasks that were on my to-do list, including getting rid of some states that controlled the visibility of components with simple boolean values. Now, I’m using a context API state called isVisible, which is an object containing all the components that need a boolean value to manage their visibility. This change simplifies the logic and makes it easier to manage visibility for modals, forms, sidebars, and anything else that pops up in the app.

Additionally, I refactored the board game form. I created a global close animation for modals and forms. Initially, when these components appeared, they had a simple jump-in animation, which was cool, but I needed to ensure they had a smooth exit as well. So, I implemented a logic to handle their visibility more gracefully when closing them. Instead of local animations, everything is now handled globally. So whenever a modal or form is closed, they will all have this jump-out animation, creating a consistent user experience. Overall, these changes should lead to cleaner code and a more polished interface.