delete twilio api files and dependencies, switch to telegram api
Removed Twilio API files and dependencies from the project as I am no longer using WhatsApp for SMS sending. Instead, I have integrated Telegram as the messaging service since it is free. The only code change made is in the service class for sending messages, which now connects to the Telegram API instead of Twilio.